git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1166

+++ src/Core/Camera.cpp
@@ -127,7 +127,7 @@ namespace TikiEngine
 		{
 			matrices.ProjectionMatrix = Matrix::Transpose(
 				Matrix::CreatePerspectiveFieldOfView(
-					MATH_PI / 4,
+					(float)(MATH_PI / 4.0),
 					(float)args.CurrentViewPort->Width / args.CurrentViewPort->Height,
 					0.01f,
 					1000.0f
